requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.1
asyncio==3.4.3
attrs==21.2.0
backcall==0.2.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
decorator==5.1.0
fast-bitrix24==1.5.1
frozenlist==1.2.0
google-api-core==2.3.2
google-api-python-client==2.33.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-bigquery==2.31.0
google-cloud-bigquery-storage==2.10.1
google-cloud-core==2.2.1
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
grpcio==1.43.0
grpcio-status==1.43.0
httplib2==0.20.2
idna==3.3
jedi==0.18.1
libcst==0.3.23
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
more-itertools==8.12.0
multidict==5.2.0
mypy-extensions==0.4.3
numpy==1.21.5
oauthlib==3.1.1
packaging==21.3
pandas==1.3.5
pandas-gbq==0.16.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.24
proto-plus==1.19.8
protobuf==3.19.1
ptyprocess==0.7.0
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydata-google-auth==1.3.0
Pygments==2.10.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
tqdm==4.62.3
traitlets==5.1.1
typing-extensions==4.0.1
typing-inspect==0.7.1
uritemplate==4.1.1
urllib3==1.26.7
wcwidth==0.2.5
yarl==1.7.2
